Initial commit of my custom MATLAB grammar for Atom.
This grammar is written in CoffeeScript that is subsequently converted
to a single CSON file. After working some with the original grammar
(from which this project is forked), I decided that it would be far
easier to do this with the support of a full programming language. The
files being committed here are thus a complete rewrite of the entire
grammar. As of now, it's working very well...